Move the various catalogue Link components into @weco/catalogue
