+++ catalogue/webapp/components/ArchiveBreadcrumb/ArchiveBreadcrumb.tsx
@@ -5,7 +5,7 @@ import Icon from '@weco/common/views/components/Icon/Icon';
 import WorkTitle from '../WorkTitle/WorkTitle';
 import { getArchiveAncestorArray } from '../../utils/works';
 import { FunctionComponent, ReactNode, useContext } from 'react';
-import WorkLink from '@weco/common/views/components/WorkLink/WorkLink';
+import WorkLink from '../WorkLink';
 import IsArchiveContext from '../IsArchiveContext/IsArchiveContext';
 import { archive, folder } from '@weco/common/icons';
