+++ catalogue/webapp/components/IIIFViewer/NoScriptViewer.tsx
@@ -14,7 +14,7 @@ import Control from '@weco/common/views/components/Buttons/Control/Control';
 import IIIFCanvasThumbnail from './IIIFCanvasThumbnail';
 import { TransformedCanvas } from '../../types/manifest';
 import { FunctionComponent } from 'react';
-import { toLink as itemLink } from '@weco/common/views/components/ItemLink/ItemLink';
+import { toLink as itemLink } from '../ItemLink';
 import { arrow } from '@weco/common/icons';
 
 const NoScriptViewerEl = styled.div`
