+++ catalogue/webapp/components/ItemLink/index.tsx
@@ -9,8 +9,8 @@ import {
   LinkFrom,
   maybeNumberCodec,
   stringCodec,
-} from '../../../utils/routes';
-import { LinkProps } from '../../../model/link-props';
+} from '@weco/common/utils/routes';
+import { LinkProps } from '@weco/common/model/link-props';
 
 const itemPropsSources = [
   'work',
