+++ catalogue/webapp/components/WorkDetails/WorkDetails.tsx
@@ -1,8 +1,8 @@
 import NextLink from 'next/link';
 import { FunctionComponent, useContext } from 'react';
 import { font } from '@weco/common/utils/classnames';
-import { toLink as worksLink } from '@weco/common/views/components/WorksLink/WorksLink';
-import { toLink as imagesLink } from '@weco/common/views/components/ImagesLink/ImagesLink';
+import { toLink as worksLink } from '../WorksLink';
+import { toLink as imagesLink } from '../ImagesLink';
 import {
   getDigitalLocationInfo,
   getDigitalLocationOfType,
@@ -25,7 +25,7 @@ import VideoPlayer from '../VideoPlayer/VideoPlayer';
 import AudioList from '../AudioList/AudioList';
 import ButtonSolidLink from '@weco/common/views/components/ButtonSolidLink/ButtonSolidLink';
 import ExplanatoryText from './ExplanatoryText';
-import { toLink as itemLink } from '@weco/common/views/components/ItemLink/ItemLink';
+import { toLink as itemLink } from '../ItemLink';
 import { trackGaEvent } from '@weco/common/utils/ga';
 import PhysicalItems from '../PhysicalItems/PhysicalItems';
 import Layout12 from '@weco/common/views/components/Layout12/Layout12';
