+++ catalogue/webapp/components/WorkTabbedNav/WorkTabbedNav.tsx
@@ -2,7 +2,7 @@ import { Work } from '@weco/catalogue/services/catalogue/types';
 import { FunctionComponent } from 'react';
 import SubNavigation from '@weco/common/views/components/SubNavigation/SubNavigation';
 import { listView, eye } from '@weco/common/icons';
-import { toLink as itemLink } from '@weco/common/views/components/ItemLink/ItemLink';
+import { toLink as itemLink } from '../ItemLink';
 
 type Props = {
   work: Work;
