+++ catalogue/webapp/services/catalogue/images.ts
@@ -20,7 +20,7 @@ import {
   emptyImagesProps,
   ImagesProps,
   toQuery,
-} from '@weco/common/views/components/ImagesLink/ImagesLink';
+} from '@weco/catalogue/components/ImagesLink';
 
 type ImageInclude =
   | 'visuallySimilar'
